+++ examples/demo/src/routes/ProductPage.tsx
@@ -1,7 +1,9 @@
 import algoliarecommend from '@algolia/recommend';
 import {
   FrequentlyBoughtTogether,
+  Recommend,
   RelatedProducts,
+  LookingSimilar,
 } from '@algolia/recommend-react';
 import { HorizontalSlider } from '@algolia/ui-components-horizontal-slider-react';
 import React from 'react';
@@ -25,7 +27,7 @@ export const ProductPage: React.FC = () => {
   }
 
   return (
-    <div>
+    <Recommend recommendClient={recommendClient}>
       <div style={{ padding: '1rem 0' }}>
         <div
           className="Hit"
@@ -47,8 +49,21 @@ export const ProductPage: React.FC = () => {
           </div>
         </div>
       </div>
+      <LookingSimilar<ProductHit>
+        indexName={indexName}
+        objectIDs={[selectedProduct.objectID]}
+        itemComponent={({ item }) => (
+          <Hit hit={item} insights={insights} onSelect={setSelectedProduct} />
+        )}
+        maxRecommendations={10}
+        threshold={75}
+        view={HorizontalSlider}
+        queryParameters={{
+          analytics: true,
+          clickAnalytics: true,
+        }}
+      />
       <FrequentlyBoughtTogether<ProductHit>
-        recommendClient={recommendClient}
         indexName={indexName}
         objectIDs={[selectedProduct.objectID]}
         itemComponent={({ item }) => (
@@ -75,7 +90,6 @@ export const ProductPage: React.FC = () => {
         )}
         fallbackComponent={() => (
           <RelatedProducts<ProductHit>
-            recommendClient={recommendClient}
             indexName={indexName}
             objectIDs={[selectedProduct.objectID]}
             itemComponent={({ item }) => (
@@ -106,7 +120,6 @@ export const ProductPage: React.FC = () => {
         )}
       />
       <RelatedProducts<ProductHit>
-        recommendClient={recommendClient}
         indexName={indexName}
         objectIDs={[selectedProduct.objectID]}
         itemComponent={({ item }) => (
@@ -131,6 +144,6 @@ export const ProductPage: React.FC = () => {
           ],
         }}
       />
-    </div>
+    </Recommend>
   );
 };