+++ packages/recommend-react/src/RelatedProducts.tsx
@@ -5,6 +5,7 @@ import {
 } from '@algolia/recommend-vdom';
 import React, { createElement, Fragment } from 'react';
 
+import { OptionalRecommendClient } from './types/OptionalRecommendClient';
 import { useRelatedProducts } from './useRelatedProducts';
 
 const UncontrolledRelatedProducts = createRelatedProductsComponent({
@@ -12,8 +13,15 @@ const UncontrolledRelatedProducts = createRelatedProductsComponent({
   Fragment,
 });
 
-export type RelatedProductsProps<TObject> = GetRelatedProductsProps<TObject> &
-  Omit<RelatedProductsVDOMProps<TObject>, 'items' | 'status'>;
+export type UseRelatedProductsProps<TObject> = OptionalRecommendClient<
+  GetRelatedProductsProps<TObject>
+>;
+
+export type RelatedProductsProps<TObject> = UseRelatedProductsProps<TObject> &
+  Omit<
+    RelatedProductsVDOMProps<TObject>,
+    'items' | 'status' | 'createElement' | 'Fragment'
+  >;
 
 export function RelatedProducts<TObject>(props: RelatedProductsProps<TObject>) {
   const { recommendations, status } = useRelatedProducts<TObject>(props);