+++ packages/recommend-react/src/TrendingItems.tsx
@@ -5,6 +5,7 @@ import {
 } from '@algolia/recommend-vdom';
 import React, { createElement, Fragment } from 'react';
 
+import { OptionalRecommendClient } from './types/OptionalRecommendClient';
 import { useTrendingItems } from './useTrendingItems';
 
 const UncontrolledTrendingItems = createTrendingItemsComponent({
@@ -12,8 +13,15 @@ const UncontrolledTrendingItems = createTrendingItemsComponent({
   Fragment,
 });
 
-export type TrendingItemsProps<TObject> = GetTrendingItemsProps<TObject> &
-  Omit<TrendingItemsVDOMProps<TObject>, 'items' | 'status'>;
+export type UseTrendingItemsProps<TObject> = OptionalRecommendClient<
+  GetTrendingItemsProps<TObject>
+>;
+
+export type TrendingItemsProps<TObject> = UseTrendingItemsProps<TObject> &
+  Omit<
+    TrendingItemsVDOMProps<TObject>,
+    'items' | 'status' | 'createElement' | 'Fragment'
+  >;
 
 export function TrendingItems<TObject>(props: TrendingItemsProps<TObject>) {
   const { recommendations, status } = useTrendingItems<TObject>(props);