+++ packages/recommend-react/src/useFrequentlyBoughtTogether.ts
@@ -1,18 +1,16 @@
+import { RecommendationsQuery } from '@algolia/recommend';
 import {
   getFrequentlyBoughtTogether,
-  GetFrequentlyBoughtTogetherProps,
   GetRecommendationsResult,
 } from '@algolia/recommend-core';
 import { useEffect, useRef, useState } from 'react';
 
+import { UseFrequentlyBoughtTogetherProps } from './FrequentlyBoughtTogether';
+import { useRecommendContext, useRecommendClient } from './RecommendContext';
 import { useAlgoliaAgent } from './useAlgoliaAgent';
 import { useStableValue } from './useStableValue';
 import { useStatus } from './useStatus';
 import { useAsyncError } from './utils/useAsyncError';
-
-export type UseFrequentlyBoughtTogetherProps<
-  TObject
-> = GetFrequentlyBoughtTogetherProps<TObject>;
 
 export function useFrequentlyBoughtTogether<TObject>({
   indexName,
@@ -21,7 +19,7 @@ export function useFrequentlyBoughtTogether<TObject>({
   queryParameters: userQueryParameters,
   recommendClient,
   threshold,
-  transformItems: userTransformItems,
+  transformItems: userTransformItems = (x) => x,
 }: UseFrequentlyBoughtTogetherProps<TObject>) {
   const throwAsyncError = useAsyncError();
   const [result, setResult] = useState<GetRecommendationsResult<TObject>>({
@@ -31,7 +29,10 @@ export function useFrequentlyBoughtTogether<TObject>({
   const objectIDs = useStableValue(userObjectIDs);
   const queryParameters = useStableValue(userQueryParameters);
 
-  useAlgoliaAgent({ recommendClient });
+  const { hasProvider, register } = useRecommendContext();
+  const { client, isContextClient } = useRecommendClient(recommendClient);
+
+  useAlgoliaAgent({ recommendClient: client });
 
   const transformItemsRef = useRef(userTransformItems);
   useEffect(() => {
@@ -39,30 +40,71 @@ export function useFrequentlyBoughtTogether<TObject>({
   }, [userTransformItems]);
 
   useEffect(() => {
-    setStatus('loading');
-    getFrequentlyBoughtTogether({
+    const param = {
       indexName,
       maxRecommendations,
       objectIDs,
       queryParameters,
-      recommendClient,
       threshold,
       transformItems: transformItemsRef.current,
+    };
+
+    if (hasProvider && isContextClient) {
+      const key = JSON.stringify(param);
+      const queries = objectIDs.map(
+        (objectID: string): RecommendationsQuery => ({
+          indexName,
+          model: 'bought-together',
+          threshold,
+          maxRecommendations,
+          objectID,
+          queryParameters,
+        })
+      );
+      return register({
+        key,
+        getParameters() {
+          return {
+            queries,
+            keyPair: {
+              key,
+              value: objectIDs.length,
+            },
+          };
+        },
+        onRequest() {
+          setStatus('loading');
+        },
+        onResult(response) {
+          setResult(response as GetRecommendationsResult<TObject>);
+          setStatus('idle');
+        },
+      });
+    }
+
+    setStatus('loading');
+    getFrequentlyBoughtTogether({
+      ...param,
+      recommendClient: client,
     })
       .then((response) => {
         setResult(response);
         setStatus('idle');
       })
       .catch(throwAsyncError);
+    return () => {};
   }, [
     indexName,
     maxRecommendations,
     objectIDs,
     queryParameters,
-    recommendClient,
+    client,
     setStatus,
     threshold,
     throwAsyncError,
+    hasProvider,
+    isContextClient,
+    register,
   ]);
 
   return {