+++ packages/recommend-react/src/useRelatedProducts.ts
@@ -1,16 +1,16 @@
+import { RecommendationsQuery } from '@algolia/recommend';
 import {
   getRelatedProducts,
-  GetRelatedProductsProps,
   GetRecommendationsResult,
 } from '@algolia/recommend-core';
 import { useEffect, useRef, useState } from 'react';
 
+import { useRecommendContext, useRecommendClient } from './RecommendContext';
+import { UseRelatedProductsProps } from './RelatedProducts';
 import { useAlgoliaAgent } from './useAlgoliaAgent';
 import { useStableValue } from './useStableValue';
 import { useStatus } from './useStatus';
 import { useAsyncError } from './utils/useAsyncError';
-
-export type UseRelatedProductsProps<TObject> = GetRelatedProductsProps<TObject>;
 
 export function useRelatedProducts<TObject>({
   fallbackParameters: userFallbackParameters,
@@ -20,7 +20,7 @@ export function useRelatedProducts<TObject>({
   queryParameters: userQueryParameters,
   recommendClient,
   threshold,
-  transformItems: userTransformItems,
+  transformItems: userTransformItems = (x) => x,
 }: UseRelatedProductsProps<TObject>) {
   const throwAsyncError = useAsyncError();
   const [result, setResult] = useState<GetRecommendationsResult<TObject>>({
@@ -31,7 +31,10 @@ export function useRelatedProducts<TObject>({
   const queryParameters = useStableValue(userQueryParameters);
   const fallbackParameters = useStableValue(userFallbackParameters);
 
-  useAlgoliaAgent({ recommendClient });
+  const { hasProvider, register } = useRecommendContext();
+  const { client, isContextClient } = useRecommendClient(recommendClient);
+
+  useAlgoliaAgent({ recommendClient: client });
 
   const transformItemsRef = useRef(userTransformItems);
   useEffect(() => {
@@ -39,29 +42,71 @@ export function useRelatedProducts<TObject>({
   }, [userTransformItems]);
 
   useEffect(() => {
-    setStatus('loading');
-    getRelatedProducts({
+    const param = {
       fallbackParameters,
       indexName,
       maxRecommendations,
       objectIDs,
       queryParameters,
-      recommendClient,
       threshold,
       transformItems: transformItemsRef.current,
+    };
+
+    if (hasProvider && isContextClient) {
+      const key = JSON.stringify(param);
+      const queries = objectIDs.map(
+        (objectID: string): RecommendationsQuery => ({
+          indexName,
+          model: 'related-products',
+          threshold,
+          maxRecommendations,
+          objectID,
+          queryParameters,
+          fallbackParameters,
+        })
+      );
+      return register({
+        key,
+        getParameters() {
+          return {
+            queries,
+            keyPair: {
+              key,
+              value: objectIDs.length,
+            },
+          };
+        },
+        onRequest() {
+          setStatus('loading');
+        },
+        onResult(response) {
+          setResult(response as GetRecommendationsResult<TObject>);
+          setStatus('idle');
+        },
+      });
+    }
+
+    setStatus('loading');
+    getRelatedProducts({
+      ...param,
+      recommendClient: client,
     })
       .then((response) => {
         setResult(response);
         setStatus('idle');
       })
       .catch(throwAsyncError);
+    return () => {};
   }, [
+    client,
     fallbackParameters,
+    hasProvider,
     indexName,
+    isContextClient,
     maxRecommendations,
     objectIDs,
     queryParameters,
-    recommendClient,
+    register,
     setStatus,
     threshold,
     throwAsyncError,