+++ packages/recommend-react/src/useTrendingItems.ts
@@ -1,16 +1,16 @@
 import {
+  BatchQuery,
   getTrendingItems,
-  GetTrendingItemsProps,
   GetTrendingItemsResult,
 } from '@algolia/recommend-core';
 import { useEffect, useRef, useState } from 'react';
 
+import { useRecommendContext, useRecommendClient } from './RecommendContext';
+import { UseTrendingItemsProps } from './TrendingItems';
 import { useAlgoliaAgent } from './useAlgoliaAgent';
 import { useStableValue } from './useStableValue';
 import { useStatus } from './useStatus';
 import { useAsyncError } from './utils/useAsyncError';
-
-export type UseTrendingItemsProps<TObject> = GetTrendingItemsProps<TObject>;
 
 export function useTrendingItems<TObject>({
   fallbackParameters: userFallbackParameters,
@@ -19,7 +19,7 @@ export function useTrendingItems<TObject>({
   queryParameters: userQueryParameters,
   recommendClient,
   threshold,
-  transformItems: userTransformItems,
+  transformItems: userTransformItems = (x) => x,
   facetName,
   facetValue,
 }: UseTrendingItemsProps<TObject>) {
@@ -31,7 +31,10 @@ export function useTrendingItems<TObject>({
   const queryParameters = useStableValue(userQueryParameters);
   const fallbackParameters = useStableValue(userFallbackParameters);
 
-  useAlgoliaAgent({ recommendClient });
+  const { hasProvider, register } = useRecommendContext();
+  const { client, isContextClient } = useRecommendClient(recommendClient);
+
+  useAlgoliaAgent({ recommendClient: client });
 
   const transformItemsRef = useRef(userTransformItems);
   useEffect(() => {
@@ -39,10 +42,8 @@ export function useTrendingItems<TObject>({
   }, [userTransformItems]);
 
   useEffect(() => {
-    setStatus('loading');
-    getTrendingItems({
-      recommendClient,
-      transformItems: transformItemsRef.current,
+    const param: BatchQuery<TObject> = {
+      model: 'trending-items',
       fallbackParameters,
       indexName,
       maxRecommendations,
@@ -50,23 +51,58 @@ export function useTrendingItems<TObject>({
       threshold,
       facetName,
       facetValue,
+      transformItems: transformItemsRef.current,
+    };
+
+    if (hasProvider && isContextClient) {
+      const key = JSON.stringify(param);
+      return register({
+        key,
+        getParameters() {
+          return {
+            queries: [param],
+            keyPair: {
+              key,
+              value: 1,
+            },
+          };
+        },
+        onRequest() {
+          setStatus('loading');
+        },
+        onResult(response) {
+          setResult(response as GetTrendingItemsResult<TObject>);
+          setStatus('idle');
+        },
+      });
+    }
+
+    setStatus('loading');
+    getTrendingItems({
+      ...param,
+      recommendClient: client,
+      transformItems: transformItemsRef.current,
     })
       .then((response) => {
         setResult(response);
         setStatus('idle');
       })
       .catch(throwAsyncError);
+    return () => {};
   }, [
     fallbackParameters,
     indexName,
     maxRecommendations,
     queryParameters,
-    recommendClient,
+    client,
     setStatus,
     threshold,
     facetName,
     facetValue,
     throwAsyncError,
+    hasProvider,
+    isContextClient,
+    register,
   ]);
 
   return {