+++ packages/recommend-react/src/utils/isPresent.ts
@@ -1,0 +1,5 @@
+export function isPresent<TObject>(
+  value: TObject
+): value is NonNullable<TObject> {
+  return value !== null && value !== undefined;
+}